More symbols/scripts renamed based on US1.0's debug
