+++ symbol_addrs.txt
@@ -490,3 +490,36 @@ gCurrentStage = 0x80174874; // type:symbol size:0x4
 aspMainTextStart = 0x800ef6c0; // type:symbol size:0xe20 rom:0xcaac0
 aspMainDataStart = 0x801119d0; // type:symbol size:0x2c0 rom:0xecdd0
 gspFast3DDataStart = 0x801111d0; // type:symbol size:0x800 rom:0xec5d0
+DMAStruct_Print = 0x800a7988; // type:function rom:0x82d88
+SetViewArea = 0x800af260; // type:function rom:0x8a660
+IsntPickup = 0x800b2260; // type:function rom:0x8d660
+checkRoomItem = 0x800b4bdc; // type:function rom:0x8ffdc
+EraseRoomItem = 0x800b4db4; // type:function rom:0x901b4
+RegistField = 0x800be9dc; // type:function rom:0x99ddc
+EraseField = 0x800bf1bc; // type:function rom:0x9a5bc
+RegistDoor = 0x800bf2e8; // type:function rom:0x9a6e8
+RegistSwitchArea = 0x800bf374; // type:function rom:0x9a774
+InitFieldSub = 0x800bf928; // type:function rom:0x9ad28
+CalcRoomInfo = 0x800bfdf4; // type:function rom:0x9b1f4
+CalcDoorInfo = 0x800c04f0; // type:function rom:0x9b8f0
+InitFieldSubScroll = 0x800c1810; // type:function rom:0x9cc10
+InitField = 0x800c2fa0; // type:function rom:0x9e3a0
+MoveField = 0x800c325c; // type:function rom:0x9e65c
+Random = 0x800c8900; // type:function rom:0xa3d00
+isPickup = 0x800b2288; // type:function rom:0x8d688
+EraseToungeEatEnemy = 0x800be580; // type:function rom:0x99980
+ChameleonFromDoor = 0x800c09ac; // type:function rom:0x9bdac
+enterBossRoom = 0x800c2f48; // type:function rom:0x9e348
+CountSpatActors = 0x800b0764; // type:function rom:0x8bb64
+IsNearPoint = 0x800d8030; // type:function rom:0xb3430
+LimitInt = 0x800b21f0; // type:function rom:0x8d5f0
+GetDirectionName = 0x800c04d4; // type:function rom:0x9b8d4
+isChange = 0x8020d890; // type:symbol size:0x4
+prevZone = 0x8020d898; // type:symbol size:0x4
+isFirstZone = 0x8020d894; // type:symbol size:0x4
+nextZone = 0x8020d89c; // type:symbol size:0x4
+VertextBufferCount = 0x8020d854; // type:symbol size:0x4
+TriangleBufferCount = 0x8020d85c; // type:symbol size:0x4
+ModelBufferCount = 0x8020d860; // type:symbol size:0x4
+CalcNextPosition = 0x800d0708; // type:function rom:0xabb08
+ClearPolygon = 0x800c8f00; // type:function rom:0xa4300
